django_th/pyproject.toml
[project]
name = "django-th"
version = "0.1.0"
description = ""
authors = [
    {name = "nPhuongNguyen",email = "phuong0909tt@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=5.2.1,<6.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "mysqlclient (>=2.2.7,<3.0.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "drf-yasg (>=1.21.10,<2.0.0)",
    "djangorestframework-xml (>=2.0.0,<3.0.0)",
]

[tool.poetry]
packages = [{include = "django_th", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
